% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggFunctions.R
\name{andList}
\alias{andList}
\title{Obtaining an AND query syntax to use two and more query elements}
\usage{
andList(query_prep = c())
}
\arguments{
\item{query_prep}{ne or more query vectors created by other query functions.
if query vectors are more than one than \code{c()} should be used for lising within the function.}
}
\value{
A list of several query elements corresponding to a particular data table
}
\description{
This function retuns a list of a comination two or more query elements.
}
\examples{
\dontrun{# to subset real-time data with the constraint of a time range and a dyad
t <- returnTimes("phoenix_rt", "20171015", "20171215")
dyad <- returnDyad('Phoexnix_rt', 'RUS', 'SYR')
# building an AND-query of a certain time and the country dyad
and_query <- andList(c(t,dyad))}
}
