% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citeData.R
\name{citeData}
\alias{citeData}
\title{Properly citing Data tables in the UTD Event data server}
\usage{
citeData(table_name = "")
}
\arguments{
\item{table_name}{a name of a data table a user specifies. Your input is NOT
case-sensitive.}
}
\value{
The texts of propor citations for publication in text and Bibtex formats
}
\description{
This function returns the citation texts of data sets a user extracts
by the UTDEventData package with a data table name.
}
\examples{
# citations for Cline Phoenix Event data
citeData(table_name = "cline_Phoenix_swb")
# citations for ICEWS or UTD real-time data
citeData(table_name = "Phoenix_rt")
}
