% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.R
\name{f3}
\alias{f3}
\title{A hiden function for relabeling query strings of one data format to the other}
\usage{
f3()
}
\description{
A hiden function for relabeling query strings of one data format to the other
}
\keyword{internal}
