% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggFunctions.R
\name{getQuerysize}
\alias{getQuerysize}
\title{Calculating a size of data a user wants to extract}
\usage{
getQuerysize(api_key = "", table_name = "", query = list())
}
\arguments{
\item{api_key}{An API key provided by a server manager at UTD}

\item{table_name}{A name of a data table a user specifies. Your input is NOT
case-sensitive.}

\item{query}{List of queries a user builds with othter aggretation functions.}
}
\value{
A data size in bytes
}
\description{
This function retruns a data size in a string format
}
\examples{
getQuerysize(api_key,"Phoenix_rt", query_list)
}
