% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.R
\name{loadFormat}
\alias{loadFormat}
\title{A hiden function for relabeling query strings of one data format to the other}
\usage{
loadFormat(format)
}
\arguments{
\item{format}{a format returned}
}
\description{
A hiden function for relabeling query strings of one data format to the other
}
\keyword{internal}
