% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previewData.R
\name{previewData}
\alias{previewData}
\title{Previewing a data structure and its variables which a user requests}
\usage{
previewData(utd_api_key = "", table_name = "")
}
\arguments{
\item{utd_api_key}{An API key provided by a server manager at UTD}

\item{table_name}{A name of a data table a user specifies. Your input is NOT
case-sensitive.}
}
\value{
A data frame of a specified data table.
}
\description{
This function retruns the e00 observations of a particular data table
}
\note{
The retrived data are invoked in a spreadsheet-style by `View()`, but the class of the variable, `_id`,
in the data is a datafame, so it returns error if you view them by `edit()`. To check the data
with `edit()`, the variable of `_id` needs to be dropped.
}
\examples{
\dontrun{ data <- previewData(utd_api_key, "icews")
View(data) # the 100 observations of ICEWS data are shown }
}
