% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.R
\name{relabel}
\alias{relabel}
\title{A Hiden function for relabeling query strings of one data format to the other}
\usage{
relabel(queryString, format)
}
\arguments{
\item{format}{a format returned to}
}
\description{
A Hiden function for relabeling query strings of one data format to the other
}
\author{
Micheal J. Shoemate
}
