% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggFunctions.R
\name{returnLatLon}
\alias{returnLatLon}
\title{Creating the location query block with the longitudes and lattidues}
\usage{
returnLatLon(lat1, lat2, lon1, lon2)
}
\arguments{
\item{lat1}{the minimum value of lattitude of a geo-location boundary}

\item{lat2}{the maximum value of latitidue of a geo-location boundary}

\item{lon1}{the minimum value of longitude of a geo-location boundary}

\item{lon2}{the maximum value of longitude of a geo-location boundary}
}
\value{
A list of geo-location query blocks
}
\description{
This function returns a list of a query block of geo-locations
}
\examples{
\dontrun{# to build the location constraint with the ranges of latitudes the longitudes
geo <- returnLatLon(-10,30,40,70)}
}
