% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggFunctions.R
\name{returnRegExp}
\alias{returnRegExp}
\title{Obtaining a list of query that use the variables in a particular data table}
\usage{
returnRegExp(api_key = "", table_name = "", pattern = "", field = "")
}
\arguments{
\item{api_key}{An API key provided by a UTD server manager}

\item{table_name}{A name of a data table a user specifies. Your input is NOT
case-sensitive.}

\item{pattern}{A pattern or feature of a specified variable}

\item{field}{A field (variable) of a data table}
}
\value{
A list of the query block with variable's features
}
\description{
This function retuns general query syntax for sepcfied data table
}
\examples{
# when a user wants to get all source actors related to governments in ICEWS
f <- returnRegExp( k, "icews","GOV","Source Name")

}
