% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggFunctions.R
\name{returnTimes}
\alias{returnTimes}
\title{Creating of a query block of a time range before applying the query to `sendQuery()`}
\usage{
returnTimes(table_name = " ", start = " ", end = " ")
}
\arguments{
\item{table_name}{A name of a data table a user specifies. Your input is NOT
case-sensitive.}

\item{start}{The "YYYYMMDD" format of the first date of a data set}

\item{end}{The "YYYYMMDD" format of the last date of a data set}
}
\value{
A list of dates for start and end of a time range
}
\description{
This function returns the list of a time range a user specifies
}
\examples{
\dontrun{# to create the time ragne between Nov. 2, 2017 and
# Nov. 4, 2017 for the subset of real-time Phoenix data
t <- returnTimes("phoenix_rt", "20171102","20171104")}
}
