% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggFunctions.R
\name{sendQuery}
\alias{sendQuery}
\title{Sending queries to the API server in order to retrieve the data set}
\usage{
sendQuery(api_key = "", table_name = "", query = list(),
  citation = TRUE)
}
\arguments{
\item{api_key}{An API key provided by a server manager at UTD}

\item{table_name}{A name of a data table a user specifies. Your input is NOT
case-sensitive.}

\item{query}{A list of query blocks a user builds with other query block functions
If you enter `query = "entire"`, then the entire data will be returned.
Please note the size of the data you request. It may cause a memory issue on your device.}

\item{citation}{The option for printing a package citation at the end of data retrival. The default of option is TRUE.}
}
\value{
A list of data and citation texts
}
\description{
This function retruns the data a user requests by a composition of query blocks
}
\examples{
sendQuery(api_key,"icews", query_block, citation = FALSE)
# To retrieve the entire data of phoenix_rt
data <- sendQuery(api_key, "phoenix_rt", "entire", citation = FALSE)
}
