% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{tableVar}
\alias{tableVar}
\title{Search variables in the specified data table}
\usage{
tableVar(table = "table_name", word = " ")
}
\arguments{
\item{table=}{the string of data table name such as 'Pheonix'}
}
\value{
All names of variable a data table has
}
\description{
Searching the list of variablees in the data table a user specifies. It returns the list of variables.
}
\examples{
> tableVar(table="Pheonix")
[1] "code"        "src_actor"       "month"    "tgt_actor"   "country_code"
[6] "year"        "id"              "source"   "date8"       "src_agent"
[11] "latitude"   "src_other_agent" "geoname"  "quad_class"  "source_text"
[16] "root_code"  "tgt_other_agent" "day"      "target"      "goldstein"
[21] "tgt_agent"  "longitude"       "url"      "_id"
}
