% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggFunctions.R
\name{andList}
\alias{andList}
\title{Obtaining an AND syntax to use two and more query lists}
\usage{
andList(query_prep = list())
}
\arguments{
\item{query_prep}{A list of other queris stored}
}
\value{
AND query syntax
}
\description{
This function retuns the syntax to combine several query lists
}
\examples{
# If you want to subset with two or more stored query blocks such as q or f,
or_query <- orList(list(q,f))

}
