% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggFunctions.R
\name{andList}
\alias{andList}
\title{Obtaining an AND query syntax to use two and more query blocks}
\usage{
andList(query_prep = list())
}
\arguments{
\item{query_prep}{A list of other query blocks that should be entered in `list()`}
}
\value{
A list of several query blocks corresponding to a particular data table
}
\description{
This function retuns the list of comination two or more query blocks.
}
\examples{
# If you want to subset with two or more stored query blocks such as q or f
and_query <- andList(list(q,f))

}
