% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citeData.R
\name{citeData}
\alias{citeData}
\title{Obtaining the text or Bibtax to properly cite the data downloaded from the UTD Event data server for publications}
\usage{
citeData(table_name = "")
}
\arguments{
\item{table_name}{a name of a data table a user specifies. Your input is NOT
case-sensitive.}
}
\value{
The texts of proper citations for publication in text and Bibtex formats
}
\description{
This function returns the citation texts in text and Bibtex formats of data sets a user extracts
by the UTDEventData package.
}
\note{
The citing texts for the package can be found after requesting data with `pullData()` or 'sendQeury()`.
}
\examples{
\\donrun{# citations for Cline Phoenix Event data
citeData("cline_Phoenix_swb")
# citations for ICEWS
citeData("icews")
# citation for real-time phoenix data
citeData("phoenix_rt")}
}
