% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{DataTables}
\alias{DataTables}
\title{Search data tables in the UTD Event data server}
\usage{
DataTables()
}
\value{
The name of data table the UTD server has
}
\description{
Searching data tables in the UTD server.
It returns the list of data table in the UTD server.
Currently only Pheonix is available.
}
\examples{
> DataTables( )
[1] "Pheonix" "OK-LN" "ICEWS"
}
