% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{DataTables}
\alias{DataTables}
\title{Searching data tables in the UTD Event data server}
\usage{
DataTables(api_key)
}
\value{
The name of data table the UTD server has
}
\description{
Searching data tables in the UTD server.
It returns the list of data table in the UTD server.
It requires an API key from the developer at UTD.
}
\examples{
> DataTables(api_key=" ")
}
