% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{DataTables}
\alias{DataTables}
\title{Searching data tables in the UTD event data server}
\usage{
DataTables(api_key = " ")
}
\arguments{
\item{api_key}{an API key provided by the UTD server manager}
}
\value{
The list of data table the UTD server has
}
\description{
Searching available data tables in the event data server.
It returns the list of data tables in the UTD server and requires an API key.
The API key can be obtained after filling out the form in the UTD event data sign-up website
(http://eventdata.utdallas.edu/signup). Please follow the direction in the \href{http://149.165.156.33:5002/signup}{UTD sign-up webpage}. \cr
You can also use this function through the reference class, \code{Table()}.
Please find the help document of the \code{Table()} function for more details on its usage.\cr

The details of data tables are illustrated in the linked websites. \cr
\itemize{
\item {Phoenix_RT:} {Real-time data from Oct. 2017 to today, please see more details in \href{http://openeventdata.org/}{Open Event Data Alliance}}.\cr
\item {ICEWS:} {Integrated Crisis Early Warning System from Harvard Dataerse from 1995 to Sep. 2016. Please see more details in \href{https://dataverse.harvard.edu/dataverse/icews}{ICEWS Dataverse}.}
\item {Phoenix NYT:} {accumulated from 1945 to 2005, information of Phoenix Event Data are found at \href{http://www.clinecenter.illinois.edu/data/event/phoenix/}{the Cline Center}.}
\item {Phoenix FBIS:} {accumulated from 1995 to 2004, the information of Phoenix Event Data are found at \href{http://www.clinecenter.illinois.edu/data/event/phoenix/}{the Cline Center}.}
\item {Phoenix SWB:} {accumulated from 1979 to 2015, the information of Phoenix Event Data are found at \href{http://www.clinecenter.illinois.edu/data/event/phoenix/}{the Cline Center}.}
 }
}
\examples{
DataTables(api_key=" ")
 "'PHOENIX_RT', 'CLINE_PHOENIX_SWB', 'CLINE_PHOENIX_FBIS', 'CLINE_PHOENIX_NYT', 'ICEWS'"
}
