% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggFunctions.R
\name{getQuerySize}
\alias{getQuerySize}
\title{Estimating a size of data a user requests to the API server}
\usage{
getQuerySize(api_key = "", table_name = "", query = list())
}
\arguments{
\item{api_key}{An API key provided by a server manager at UTD}

\item{table_name}{A name of a data table a user specifies. Your input is NOT
case-sensitive.}

\item{query}{A list of query blocks a user builds with other query block functions
If you enter `query = "entire"`, then the size of the entire data will be returned.}
}
\value{
A data size in bytes
}
\description{
This function retruns a data size in a string format
}
\examples{
getQuerySize(api_key,"Phoenix_rt", query_blocks)
To get the size of entire data of ICEWS
getQuerySize(api_key, "icews", "entire")
}
