% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggFunctions.R
\name{orList}
\alias{orList}
\title{Obtaining an OR syntax to use two and more query lists}
\usage{
orList(query_prep = list())
}
\arguments{
\item{query_prep}{A list of other query blocks that should be entered in `list()`}
}
\value{
A list of several query blocks corresponding to a particular data table
}
\description{
This function retuns the syntax to combine several query lists
}
\examples{
# If you want to combine stored query blocks such as q or f,
or_query <- orList(list(q,f))

}
