% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previewData.R
\name{previewData}
\alias{previewData}
\title{Checking the data frame of a particular data table}
\usage{
previewData(api_key = "", table_name = "")
}
\arguments{
\item{api_key}{An API key provided by a server manager at UTD}

\item{table_name}{A name of a data table a user specifies. Your input is NOT
case-sensitive.}
}
\value{
A data frame of a specified data table.
}
\description{
This function retruns the 100 obsertations of a particular data table at the UTD API server
}
\note{
The retrived data are shown in `View()`, but the class of the variable, `_id`, is a dataframe. To check the data
with `edit()` with no warning or error, the variable of `_id` need to be dropped.
}
\examples{
\dontrun{ data <- previewData(api_key, "icews")
View(data) # the 100 observations of ICEWS data are shown }
}
