% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullData.R
\name{pullData}
\alias{pullData}
\title{Extracting event data from the UTD real-time Event database}
\usage{
pullData(api_key = " ", table_name = "Phoenix", country = list(),
  start = " ", end = " ")
}
\arguments{
\item{api_key}{The UTD API key provided by the developer}

\item{table_name}{The name of data table you want to have. e.g. list=("USA","CAN")}

\item{country}{List of countries you want to look for with ISO ALPHA-3 Code.}

\item{start}{The "YYYYMMDD" format of the first date of a data set.}

\item{end}{The "YYYYMMDD" format of the end date of a data set.}
}
\value{
Real time event data
}
\description{
This is the main function to extract subdata from the UTD Event data server by country names and time ranges.
             The API key is required. Please contact the developer (\url{https://github.com/Sayeedsalam/spec-event-data-server}) to get the API key.
}
\examples{
pullData(api_key=" ", table_name="Phoenix", country=list("USA","MEX","SYR","CHN"),
 start="20171101", end="20171112")
}
