% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.R
\name{relabel}
\alias{relabel}
\title{A hiden function for relabeling query strings of one data format to the other}
\usage{
relabel(queryString, format)
}
\arguments{
\item{queryString}{a querying sentence of a data format}

\item{format}{a format returned}
}
\description{
A hiden function for relabeling query strings of one data format to the other
}
\author{
Michael J. Shoemate
}
\keyword{internal}
