% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggFunctions.R
\name{returnDyad}
\alias{returnDyad}
\title{Creating a list of the dayd countries of source and target actors}
\usage{
returnDyad(table_name, source, target)
}
\arguments{
\item{table_name}{a name of a data table a user specifies. Your input is NOT
case-sensitive.}

\item{target}{The name of a target country either an ISO code or a country name}

\item{soure}{The name of a source country either an ISO code or a country name}
}
\value{
A list of the queary syntax of source and target countries
}
\description{
This function returns a vector of query syntax for a dayd information
}
\examples{
 # When you have a dyad query of Syria as a source and the United State as a target
dyad <- returnDayd("pheonix_rt", "SYR", "USA")
}
