% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggFunctions.R
\name{returnLatLon}
\alias{returnLatLon}
\title{Creating a vector of location boundaries with longitudes and lattidues}
\usage{
returnLatLon(lat1, lat2, lon1, lon2)
}
\arguments{
\item{lat1}{the minimum value of lattitude of a target boundary}

\item{lat2}{the maximum value of latitidue of a target boundary}

\item{log1}{the minimum value of longitude of a target boundary}

\item{log2}{the maximum value of longitude of a target boundary}
}
\value{
A vector of queary syntax of latidues and longitudes
}
\description{
This function returns a vector of query syntax for geo-locations
}
\examples{
q <- returnLatLon(-80,30,20,100)
}
