% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggFunctions.R
\name{returnRegExp}
\alias{returnRegExp}
\title{Obtaining a regular expression for API queries based on a specified data table}
\usage{
returnRegExp(api_key = "", table_name = "", pattern = "", field = "")
}
\arguments{
\item{api_key}{An API key provided by a UTD server manager}

\item{table_name}{A name of a data table a user specifies. Your input is NOT
case-sensitive.}

\item{pattern}{A pattern of an actor in events}

\item{field}{A field of a data table}
}
\value{
JSON query syntax
}
\description{
This function retuns general query syntax for sepcfied data table
}
\examples{
# when a user wants to get all source actors related to governments
f <- returnRegExp( k, "icews","GOV","Source Name")

}
