% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggFunctions.R
\name{sendQuery}
\alias{sendQuery}
\title{Sending queries to the API server in order to retrieve the data set}
\usage{
sendQuery(api_key = "", table_name = "", query = list(),
  citation = TRUE)
}
\arguments{
\item{api_key}{An API key provided by a server manager at UTD}

\item{table_name}{A name of a data table a user specifies. Your input is NOT case-sensitive.}

\item{query}{A list of query blocks or a single query block a user builds with the other query functions.}

\item{citation}{The option for printing a package citation at the end of data retrival. The default of option is TRUE.}
}
\value{
A list with components
    \itemize{
         \item{}{\code{$data    }   a data frame of requested data. An attribute of the data can be formatted as data.frame. Please check the features of data if it's necessary}
         \item{}{\code{$citation}   a text of the package citation}
          }
}
\description{
This function retruns the data and the package citation a user requests
}
\note{
If an error message is returned, please increase the memory size your R is allocated. This error is more
frequently occurred in a Windows machine.
}
\examples{
\dontrun{ # to store the ICEWS subset in the vector of myData without the citation
myData <- sendQuery(api_key,"icews", query_block, citation = FALSE)}
}
