% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table.R
\docType{class}
\name{Table-class}
\alias{Table-class}
\alias{Table}
\title{A Reference class for applying an API key to all data extracting functions.}
\description{
A Reference class for applying an API key to all data extracting functions.
}
\section{Fields}{

\describe{
\item{\code{a}}{string of the API key from the developer at UTD}
}}


