% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table.R
\docType{class}
\name{Table-class}
\alias{Table-class}
\alias{Table}
\title{A Reference class for applying an API key to all data extracting functions.}
\description{
Once a referece class is set, a user does not need to go through an API key
to call searching functions
}
\section{Fields}{

\describe{
\item{\code{api_key}}{An API key from the developer at UTD}
}}


\examples{
>obj<-Table$new() for creating an object
>obj$setAPIKey("....")
>obj$DataTable()
}
