% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{tableVar}
\alias{tableVar}
\title{Searching the variables in a particular data table}
\usage{
tableVar(api_key = " ", table = "table_name", lword = " ")
}
\arguments{
\item{api_key}{an API key from the developer at UTD}

\item{table}{a specific data table a user want to exploer its variables}

\item{lword}{a look-up word for a particular variable name in a data table}
}
\value{
The variables list in a particular data table
}
\description{
Searching the list of variablees in the data table a user specifies. It returns the list of variables.
}
\examples{
> tableVar(api_key="...", table="phoenix_rt")
 # when searching the variables in Phoenix_RT
[1] " code", "src_actor", "month", "tgt_actor", ....
> tableVar(api_key="...", table="icews", lword="tar")
 # when searhing the variable which includes the word of "tar" in ICEWS
[1] " Target Name"    " Target Sectors", ....
}
